#291 fix state
