+++ proxy/common_neon/constants.py
@@ -1,0 +1,11 @@
+KECCAK_PROGRAM = "KeccakSecp256k11111111111111111111111111111"
+INCINERATOR_PUBKEY = "1nc1nerator11111111111111111111111111111111"
+SYSVAR_INSTRUCTION_PUBKEY = "Sysvar1nstructions1111111111111111111111111"
+
+STORAGE_SIZE = 128*1024
+
+ACCOUNT_SEED_VERSION=b'\1'
+
+COLLATERALL_POOL_MAX=10
+
+EMPTY_STORAGE_TAG=0
