+++ proxy/plugin/solana_rest_api.py
@@ -21,20 +21,20 @@ from ..http.websocket import WebsocketFrame
 from ..http.server import HttpWebServerBasePlugin, httpProtocolTypes
 from .eth_proto import Trx as EthTrx
 from solana.rpc.api import Client as SolanaClient
-from sha3 import keccak_256
 import base58
 import traceback
 import threading
 from .solana_rest_api_tools import EthereumAddress, getTokens, getAccountInfo, \
-    call_signed, call_emulated, EthereumError, neon_config_load, MINIMAL_GAS_PRICE
+    call_signed, neon_config_load
 from solana.rpc.commitment import Commitment, Confirmed
 from web3 import Web3
 import logging
 from ..core.acceptor.pool import proxy_id_glob
 import os
-from ..indexer.utils import get_trx_results, LogDB
-from ..indexer.sql_dict import SQLDict
-from proxy.environment import evm_loader_id, solana_cli, solana_url
+from proxy.indexer.utils import get_trx_results, LogDB
+from proxy.indexer.sql_dict import SQLDict
+from proxy.environment import evm_loader_id, solana_cli, solana_url, MINIMAL_GAS_PRICE
+from proxy.common_neon.emulator_interactor import EthereumError, call_emulated
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
